bot/requirements.txt
python-dotenv>=1.0
httpx>=0.27
